refactor: eliminate 3 architectural duplications across api/web
- **`mapEnrichedRun` helper (apps/api/src/services/state/runs.ts)** —
  `listRunsWithFilter`, `listGlobalRuns`, and `getRunFull` were each
  inlining the same six-field null-coalescing mapping after their
  shared `enrichedRunSelect()` / 5-JOIN read. Extracted the post-fetch
  transform into `mapEnrichedRun(row)` so the "enriched run" wire
  shape now has a single source of truth. The JOINs stay inline (the
  existing comment already explains why — Drizzle's query-builder
  types don't compose through a helper).

- **Delete `APPSTRATE_SCOPES` Proxy shim
  (apps/api/src/modules/oidc/auth/scopes.ts)** — 21-line
  `@deprecated` export wrapped in a warn-on-read Proxy, kept only
  for "external consumers" that, pre-1.0, don't exist. The platform
  itself already uses `getAppstrateScopes()` (per the shim's own
  comment). Migrated the single real caller (`discovery.test.ts`)
  to `getAppstrateScopes()` and updated README + test names.

- **Move `SmtpConfigView` / `SocialProviderView` / `SocialProviderId`
  to `@appstrate/shared-types`** — the three types were defined
  identically in backend services (`modules/oidc/services/{smtp,social}.ts`)
  and in the frontend hook (`apps/web/src/modules/oidc/hooks/use-app-auth-config.ts`).
  Backend now re-exports from shared-types so module public API
  stays unchanged; frontend imports from shared-types directly.
  Eliminates the drift risk across the HTTP contract boundary
  (`/api/applications/:id/smtp-config`,
  `/api/applications/:id/social-providers/:provider`).

Net: -37 lines, three drift sources collapsed to one. Typecheck,
eslint, prettier, OpenAPI verification, and 580/580 API unit tests
all pass.
